[Feat] #27 - ImageGridView 관련 기존 코드 삭제 및 수정
